==> 02_prep_data.ipynb <==
import pandas as pd
from pandas import ExcelWriter
from datetime import datetime
import os

industry_data = pd.read_csv('ind_niftytotalmarket_list.csv')

industry_data

industry_data['Industry'].nunique()

industry_data['Industry'].unique()

industries_list = list(industry_data['Industry'].unique())

industry_wise_dict = {}
sum = 0
for industry in industries_list:
    industry_wise_dict[industry] = list(industry_data[industry_data['Industry'] == industry]['Symbol'].unique())
    print(len(industry_wise_dict[industry]))
    sum += len(industry_wise_dict[industry])

sum

industry_wise_dict

import json

file_path = "industry_wise_stocks.json"

with open(file_path, 'w') as file:
    json.dump(industry_wise_dict, file)

df = pd.DataFrame()

for k,v in industry_wise_dict.items():
    
    if len(v) < 108:
        v = v + ['']*(108-len(v))
    
    df[k] = v

df

df.to_csv('indutry_wise_list.csv')

industries_list

industries_list_modified = industries_list.copy()

industries_list_modified = industries_list
industries_list_modified.remove('Diversified')
industries_list_modified.remove('Utilities')
industries_list_modified.remove('Forest Materials')

industries_list_modified, len(industries_list_modified)

final_stock_df = pd.read_csv('Final Stock Data.csv', index_col='Date')
final_stock_df.index = pd.to_datetime(final_stock_df.index)
final_stock_df

final_stock_df.columns[final_stock_df.isna().any()].tolist()

all_years_df_dict = {}

years = [
    ('2018','11'), ('2018','12'),

    ('2019','1'), ('2019','2'), ('2019','3'), ('2019','4'), ('2019','5'), ('2019','6'), 
    ('2019','7'), ('2019','8'), ('2019','9'), ('2019','10'), ('2019','11'), ('2019','12'),

    ('2020','1'), ('2020','2'), ('2020','3'), ('2020','4'), ('2020','5'), ('2020','6'),
    ('2020','7'), ('2020','8'), ('2020','9'), ('2020','10'), ('2020','11'), ('2020','12'),

    ('2021','1'), ('2021','2'), ('2021','3'), ('2021','4'), ('2021','5'), ('2021','6'),
    ('2021','7'), ('2021','8'), ('2021','9'), ('2021','10'), ('2021','11'), ('2021','12'),

    ('2022','1'), ('2022','2'), ('2022','3'), ('2022','4'), ('2022','5'), ('2022','6'),
    ('2022','7'), ('2022','8'), ('2022','9'), ('2022','10'), ('2022','11'), ('2022','12'),

    ('2023','1'), ('2023','2'), ('2023','3'), ('2023','4'), ('2023','5'), ('2023','6'),
    ('2023','7'), ('2023','8'), ('2023','9'), ('2023','10'), ('2023','11'), ('2023','12')
]

for year, month in years:
    year_month = year+'_'+month
    end_year_int = int(year)
    month = int(month)
    start_year_int = end_year_int-5
    start_date = datetime(start_year_int, month, 1)
    end_date = datetime(end_year_int, month, 1)
    df = final_stock_df[(final_stock_df.index >= start_date) & (final_stock_df.index < end_date)]
    all_years_df_dict[year_month] = df

len(all_years_df_dict)

def get_one_week_returns(stock_df):
    one_week_returns_all_stocks = pd.DataFrame(index=stock_df.index)

    final_symbols = list(stock_df.columns)

    for symbol in final_symbols:

        one_week_returns = []

        for i in range(len(stock_df)):
            if i+1 < len(stock_df):
                pct_change = ((stock_df[symbol].iloc[i+1] - stock_df[symbol].iloc[i])/stock_df[symbol].iloc[i])
            else:
                pct_change = 0
            one_week_returns.append(pct_change)

        one_week_returns_all_stocks[symbol] = one_week_returns
    
    return one_week_returns_all_stocks

all_years_df_dict['2023_12']

year_month = '2023_12'

directory_name = f'./Industry Wise Data/{year_month}'
os.makedirs(directory_name, exist_ok=True)

for industry in industries_list_modified:
    
    industry_list = industry_wise_dict[industry]

    modified_industry_list = [symbol + ".NS" for symbol in industry_list if symbol + ".NS" in list(all_years_df_dict[year_month].columns)]

    final_list = []
    for stock in modified_industry_list:
        final_list.append(stock)

    modified_industry_list = final_list
    print(len(modified_industry_list), year_month)
    
    if len(modified_industry_list) > 2:

        historical_price = all_years_df_dict[year_month][modified_industry_list]
        weekly_returns = get_one_week_returns(historical_price)
        covariance_matrix = weekly_returns.cov()

        with ExcelWriter(f'./Industry Wise Data/{year_month}/{industry}.xlsx') as writer:
            historical_price.to_excel(writer, sheet_name='historical_price', index=True)
            weekly_returns.to_excel(writer, sheet_name='weekly_returns', index=True)
            covariance_matrix.to_excel(writer, sheet_name='covariance_matrix', index=True)
