==> 05_optimize_selected_stocks.ipynb <==
import json
import pandas as pd
import statsmodels.api as sm
import yfinance as yf
from scipy.optimize import minimize
import math
import numpy as np
import warnings
warnings.filterwarnings('ignore')

# 'yyyy-mm-dd'
nifty50 = yf.download('^NSEI', '2023-01-01', '2024-01-01', interval="1d")
nifty50 = nifty50[['Adj Close']]
nifty50.columns = [['Nifty 50']]
nifty50.index = pd.to_datetime(nifty50.index)
nifty50

file_name = 'industryWise_monthWise_stocks_intersection_of_both.json'

with open(file_name, 'r') as file:
    data = json.load(file)

yearWise_dict_stocks_intersection_of_both = {}
for year, dict in data.items():
    yearWise_dict_stocks_intersection_of_both[year] = []
    for industry, list_ in dict.items():
        yearWise_dict_stocks_intersection_of_both[year] += list_

yearWise_dict_stocks_intersection_of_both

file_name = 'industryWise_monthWise_stocks_no_zero_weightage.json'

with open(file_name, 'r') as file:
    data = json.load(file)

yearWise_dict_stocks_no_zero_weightage = {}
for year, dict in data.items():
    yearWise_dict_stocks_no_zero_weightage[year] = []
    for industry, list_ in dict.items():
        yearWise_dict_stocks_no_zero_weightage[year] += list_

yearWise_dict_stocks_no_zero_weightage

file_name = 'industryWise_monthWise_stocks_overall_highest_weightage.json'

with open(file_name, 'r') as file:
    data = json.load(file)

yearWise_dict_stocks_overall_highest_weightage = {}
for year, dict in data.items():
    yearWise_dict_stocks_overall_highest_weightage[year] = []
    for industry, list_ in dict.items():
        yearWise_dict_stocks_overall_highest_weightage[year] += list_

yearWise_dict_stocks_overall_highest_weightage

final_stock_df = pd.read_csv('Final Stock Data.csv', index_col='Date')
final_stock_df.index = pd.to_datetime(final_stock_df.index)
final_stock_df

if len(final_stock_df) > len(nifty50):
    final_stock_df = final_stock_df[final_stock_df.index.isin(nifty50.index)]
elif len(nifty50) > len(final_stock_df):
    nifty50 = nifty50[nifty50.index.isin(final_stock_df.index)]

final_stock_df = final_stock_df.backfill()
nifty50 = nifty50.backfill()

# Function to perform linear regression and return alpha and beta
def regress_stock_with_index(df, stock_column, nifty50_temp_wr):
    X = nifty50_temp_wr['Nifty 50']
    Y = df[stock_column]
    X = sm.add_constant(X) # adding a constant for intercept
    model = sm.OLS(Y, X).fit()
    return model.params

# Single Index Model - Minimize Variance

def smi_min_variance(index_average_return, index_std_dev, alpha_list, beta_list, expected_returns, num_symbs):

    def get_alpha_beta_total(weights):
        alpha_total = 0
        i = 0
        for alpha in alpha_list:
            alpha_total += alpha*weights[i]
            i += 1

        i = 0
        beta_total = 0
        for beta in beta_list:
            beta_total += beta*weights[i]
            i += 1

        return alpha_total, beta_total

    def daily_pf_rtn(weights):
        return get_alpha_beta_total(weights)[0] + (get_alpha_beta_total(weights)[1] * index_average_return)

    def ann_pf_rtn(weights):
        return ((1+daily_pf_rtn(weights))**52)-1

    def daily_pf_var(weights):
        return (get_alpha_beta_total(weights)[1] * index_std_dev)**2

    def daily_std_dev(weights):
        return get_alpha_beta_total(weights)[1] * index_std_dev
    
    def ann_std_dev(weights):
        daily_std_dev_ = get_alpha_beta_total(weights)[1] * index_std_dev
        return daily_std_dev_*math.sqrt(52)
    
    def objective_function(weights):
        var = daily_pf_var(weights)
        return var
    
    def non_zero(weights):
        count = 0
        for w in weights:
            if round(w,4) != 0.0000:
                count += 1
        return count
    
    def no_above_25(weights):
        for w in weights:
            if round(w,4) >= 0.2500:
                return 0
        return 1

    weights = []
    for i in range(num_symbs):
        weights.append(1/num_symbs)

    # Constraints and bounds
    constraints = (
        {'type': 'eq', 'fun': lambda x: np.sum(x) - 1},
        {'type': 'ineq', 'fun': lambda x: ann_pf_rtn(x) - expected_returns},
        {'type': 'ineq', 'fun': lambda x: non_zero(x) - 8},
        {'type': 'eq', 'fun': lambda x: no_above_25(x) - 1}
    )

    bounds = []
    for i in range(num_symbs):
        bounds.append((0,1))

    # Perform the optimization using the initial guess from before
    opt_result_1 = minimize(
        objective_function,
        weights,
        method='trust-constr',
        bounds=bounds,
        constraints=constraints
    )

    opt_result_2 = minimize(
        objective_function,
        weights,
        method='slsqp',
        bounds=bounds,
        constraints=constraints
    )

    if opt_result_1.fun <= opt_result_2.fun:
        opt_result = opt_result_1
    else:
        opt_result = opt_result_2

    new_weights_temp = list(opt_result.x)
    new_weights = []
    for i in new_weights_temp:
        new_weights.append(round(i,4))

    result_dict = {
        'Optimal Weights': new_weights,
        'Min Variance': opt_result.fun,
        'Weekly Return': daily_pf_rtn(new_weights),
        'Annual Return': ann_pf_rtn(new_weights),
        'Weekly Std Dev': daily_std_dev(new_weights),
        'Annual Std Dev': ann_std_dev(new_weights)
    }
    
    return result_dict

# Single Index Model - Maximize Returns

def smi_max_returns(index_average_return, index_std_dev, alpha_list, beta_list, maximum_risk, num_symbs):

    def get_alpha_beta_total(weights):
        alpha_total = 0
        i = 0
        for alpha in alpha_list:
            alpha_total += alpha*weights[i]
            i += 1

        i = 0
        beta_total = 0
        for beta in beta_list:
            beta_total += beta*weights[i]
            i += 1

        return alpha_total, beta_total

    def daily_pf_rtn(weights):
        return get_alpha_beta_total(weights)[0] + (get_alpha_beta_total(weights)[1] * index_average_return)

    def ann_pf_rtn(weights):
        return ((1+daily_pf_rtn(weights))**52)-1

    def daily_pf_var(weights):
        return (get_alpha_beta_total(weights)[1] * index_std_dev)**2

    def daily_std_dev(weights):
        return get_alpha_beta_total(weights)[1] * index_std_dev
    
    def ann_std_dev(weights):
        daily_std_dev_ = get_alpha_beta_total(weights)[1] * index_std_dev
        return daily_std_dev_*math.sqrt(52)
    
    def objective_function(weights):
        ret = daily_pf_rtn(weights)
        return -ret
    
    def non_zero(weights):
        count = 0
        for w in weights:
            if round(w,4) != 0.0000:
                count += 1
        return count
    
    def no_above_25(weights):
        for w in weights:
            if round(w,4) >= 0.2500:
                return 0
        return 1

    weights = []
    for i in range(num_symbs):
        weights.append(1/num_symbs)

    # Constraints and bounds
    constraints = (
        {'type': 'eq', 'fun': lambda x: np.sum(x) - 1},
        {'type': 'ineq', 'fun': lambda x: maximum_risk - ann_std_dev(x)},
        {'type': 'ineq', 'fun': lambda x: non_zero(x) - 8},
        {'type': 'eq', 'fun': lambda x: no_above_25(x) - 1}
    )
    
    bounds = []
    for i in range(num_symbs):
        bounds.append((0,1))
    
    # Perform the optimization using the initial guess from before
    opt_result_1 = minimize(
        objective_function,
        weights,
        method='trust-constr',
        bounds=bounds,
        constraints=constraints
    )

    opt_result_2 = minimize(
        objective_function,
        weights,
        method='slsqp',
        bounds=bounds,
        constraints=constraints
    )

    if opt_result_1.fun <= opt_result_2.fun:
        opt_result = opt_result_1
    else:
        opt_result = opt_result_2

    new_weights_temp = list(opt_result.x)
    new_weights = []
    for i in new_weights_temp:
        new_weights.append(round(i,4))

    result_dict = {
        'Optimal Weights': new_weights,
        'Min Variance': opt_result.fun,
        'Weekly Return': daily_pf_rtn(new_weights),
        'Annual Return': ann_pf_rtn(new_weights),
        'Weekly Std Dev': daily_std_dev(new_weights),
        'Annual Std Dev': ann_std_dev(new_weights)
    }
    
    return result_dict

expected_returns = 0.2
maximum_risk = 0.4

def top_8_stocks(stock_dict):
    # Flatten the dictionary into a list of tuples
    all_stocks = []
    for key in stock_dict:
        all_stocks.extend(stock_dict[key])

    # Sort the list based on weights (second element of each tuple)
    sorted_stocks = sorted(all_stocks, key=lambda x: x[1], reverse=True)

    # Get the top 8 stocks
    top_8 = sorted_stocks[:8]

    symbs = []
    for i,j in top_8:
        symbs.append(i)

    return symbs

results_stocks_overall_highest_weightage = {}
portfolio_overall_highest_weightage = {}

for year_month, stocks in yearWise_dict_stocks_overall_highest_weightage.items():

    print(year_month)
    year = year_month[:4]
    month = year_month[5:]
    print(year, month)

    historical_price = final_stock_df[stocks]
    historical_price = historical_price[(historical_price.index >= f'{month}-01-{int(year)-5}') & (historical_price.index < f'{month}-01-{year}')]

    nifty50_temp = nifty50[(nifty50.index >= f'{month}-01-{int(year)-5}') & (nifty50.index < f'{month}-01-{year}')]
    nifty50_temp_wr = nifty50_temp.pct_change().dropna()

    weekly_returns = historical_price.pct_change().dropna()
    cov_matrix = weekly_returns.cov()

    symbols = historical_price.columns.to_list()

    results = {}
    stock_list = symbols.copy()

    for stock in stock_list:
        results[stock] = regress_stock_with_index(weekly_returns, stock, nifty50_temp_wr)

    alpha_beta_dict = {}
    alpha_list = []
    beta_list = []

    for stock in stock_list:
        alpha = results[stock]['const']
        beta = results[stock][('Nifty 50',)]
        alpha_list.append(alpha)
        beta_list.append(beta)
        alpha_beta_tuple = (alpha, beta)
        alpha_beta_dict[stock] = alpha_beta_tuple

    index_std_dev = nifty50_temp_wr.std().values[0]
    index_average_return = nifty50_temp_wr.mean().values[0]

    num_symbs = len(historical_price.columns)

    results_dict = smi_min_variance(index_average_return, index_std_dev, alpha_list, beta_list, expected_returns, num_symbs)

    results_stocks_overall_highest_weightage[year_month] = results_dict

    portfolio = []
    for i in range(len(symbols)):
        symbol = symbols[i]
        weight = results_dict['Optimal Weights'][i]
        portfolio.append((symbol, weight))

    portfolio_overall_highest_weightage[year_month] = portfolio

    symbols = top_8_stocks(portfolio_overall_highest_weightage)

    historical_price = final_stock_df[symbols]
    historical_price = historical_price[(historical_price.index >= f'{month}-01-{int(year)-5}') & (historical_price.index < f'{month}-01-{year}')]

    weekly_returns = historical_price.pct_change().dropna()
    cov_matrix = weekly_returns.cov()

    symbols = historical_price.columns.to_list()

    results = {}
    stock_list = symbols.copy()

    for stock in stock_list:
        results[stock] = regress_stock_with_index(weekly_returns, stock, nifty50_temp_wr)

    alpha_beta_dict = {}
    alpha_list = []
    beta_list = []

    for stock in stock_list:
        alpha = results[stock]['const']
        beta = results[stock][('Nifty 50',)]
        alpha_list.append(alpha)
        beta_list.append(beta)
        alpha_beta_tuple = (alpha, beta)
        alpha_beta_dict[stock] = alpha_beta_tuple

    index_std_dev = nifty50_temp_wr.std().values[0]
    index_average_return = nifty50_temp_wr.mean().values[0]

    num_symbs = len(historical_price.columns)

    results_dict = smi_max_returns(index_average_return, index_std_dev, alpha_list, beta_list, maximum_risk, num_symbs)

    results_stocks_overall_highest_weightage[year_month] = results_dict

    portfolio = []
    for i in range(len(symbols)):
        symbol = symbols[i]
        weight = results_dict['Optimal Weights'][i]
        portfolio.append((symbol, weight))

    portfolio_overall_highest_weightage[year_month] = portfolio

portfolio_overall_highest_weightage

json_file_path = 'portfolio_overall_highest_weightage.json'

with open(json_file_path, 'w') as json_file:
    json.dump(portfolio_overall_highest_weightage, json_file)

results_stocks_no_zero_weightage = {}
portfolio_no_zero_weightage = {}

for year_month, stocks in yearWise_dict_stocks_no_zero_weightage.items():

    year = year_month[:4]

    historical_price = final_stock_df[stocks]
    historical_price = historical_price[(historical_price.index >= f'{month}-01-{int(year)-5}') & (historical_price.index < f'{month}-01-{year}')]

    nifty50_temp = nifty50[(nifty50.index >= f'{month}-01-{int(year)-5}') & (nifty50.index < f'{month}-01-{year}')]
    nifty50_temp_wr = nifty50_temp.pct_change().dropna()

    weekly_returns = historical_price.pct_change().dropna()
    cov_matrix = weekly_returns.cov()

    symbols = historical_price.columns.to_list()

    results = {}
    stock_list = symbols.copy()

    for stock in stock_list:
        results[stock] = regress_stock_with_index(weekly_returns, stock, nifty50_temp_wr)

    alpha_beta_dict = {}
    alpha_list = []
    beta_list = []

    for stock in stock_list:
        alpha = results[stock]['const']
        beta = results[stock][('Nifty 50',)]
        alpha_list.append(alpha)
        beta_list.append(beta)
        alpha_beta_tuple = (alpha, beta)
        alpha_beta_dict[stock] = alpha_beta_tuple

    index_std_dev = nifty50_temp_wr.std().values[0]
    index_average_return = nifty50_temp_wr.mean().values[0]

    num_symbs = len(historical_price.columns)

    results_dict = smi_min_variance(index_average_return, index_std_dev, alpha_list, beta_list, expected_returns, num_symbs)

    results_stocks_no_zero_weightage[year_month] = results_dict

    portfolio = []
    for i in range(len(symbols)):
        symbol = symbols[i]
        weight = results_dict['Optimal Weights'][i]
        portfolio.append((symbol, weight))

    portfolio_no_zero_weightage[year_month] = portfolio

    symbols = top_8_stocks(portfolio_no_zero_weightage)

    historical_price = final_stock_df[symbols]
    historical_price = historical_price[(historical_price.index >= f'{month}-01-{int(year)-5}') & (historical_price.index < f'{month}-01-{year}')]

    weekly_returns = historical_price.pct_change().dropna()
    cov_matrix = weekly_returns.cov()

    symbols = historical_price.columns.to_list()

    results = {}
    stock_list = symbols.copy()

    for stock in stock_list:
        results[stock] = regress_stock_with_index(weekly_returns, stock, nifty50_temp_wr)

    alpha_beta_dict = {}
    alpha_list = []
    beta_list = []

    for stock in stock_list:
        alpha = results[stock]['const']
        beta = results[stock][('Nifty 50',)]
        alpha_list.append(alpha)
        beta_list.append(beta)
        alpha_beta_tuple = (alpha, beta)
        alpha_beta_dict[stock] = alpha_beta_tuple

    index_std_dev = nifty50_temp_wr.std().values[0]
    index_average_return = nifty50_temp_wr.mean().values[0]

    num_symbs = len(historical_price.columns)

    results_dict = smi_max_returns(index_average_return, index_std_dev, alpha_list, beta_list, maximum_risk, num_symbs)

    results_stocks_no_zero_weightage[year_month] = results_dict

    portfolio = []
    for i in range(len(symbols)):
        symbol = symbols[i]
        weight = results_dict['Optimal Weights'][i]
        portfolio.append((symbol, weight))

    portfolio_no_zero_weightage[year_month] = portfolio

portfolio_no_zero_weightage

json_file_path = 'portfolio_no_zero_weightage.json'

with open(json_file_path, 'w') as json_file:
    json.dump(portfolio_no_zero_weightage, json_file)

results_stocks_intersection_of_both = {}
portfolio_intersection_of_both = {}

for year_month, stocks in yearWise_dict_stocks_intersection_of_both.items():

    year = year_month[:4]

    historical_price = final_stock_df[stocks]
    historical_price = historical_price[(historical_price.index >= f'{month}-01-{int(year)-5}') & (historical_price.index < f'{month}-01-{year}')]

    nifty50_temp = nifty50[(nifty50.index >= f'{month}-01-{int(year)-5}') & (nifty50.index < f'{month}-01-{year}')]
    nifty50_temp_wr = nifty50_temp.pct_change().dropna()

    weekly_returns = historical_price.pct_change().dropna()
    cov_matrix = weekly_returns.cov()

    symbols = historical_price.columns.to_list()

    results = {}
    stock_list = symbols.copy()

    for stock in stock_list:
        results[stock] = regress_stock_with_index(weekly_returns, stock, nifty50_temp_wr)

    alpha_beta_dict = {}
    alpha_list = []
    beta_list = []

    for stock in stock_list:
        alpha = results[stock]['const']
        beta = results[stock][('Nifty 50',)]
        alpha_list.append(alpha)
        beta_list.append(beta)
        alpha_beta_tuple = (alpha, beta)
        alpha_beta_dict[stock] = alpha_beta_tuple

    index_std_dev = nifty50_temp_wr.std().values[0]
    index_average_return = nifty50_temp_wr.mean().values[0]

    num_symbs = len(historical_price.columns)

    results_dict = smi_min_variance(index_average_return, index_std_dev, alpha_list, beta_list, expected_returns, num_symbs)

    results_stocks_intersection_of_both[year_month] = results_dict

    portfolio = []
    for i in range(len(symbols)):
        symbol = symbols[i]
        weight = results_dict['Optimal Weights'][i]
        portfolio.append((symbol, weight))

    portfolio_intersection_of_both[year_month] = portfolio

    symbols = top_8_stocks(portfolio_intersection_of_both)

    historical_price = final_stock_df[symbols]
    historical_price = historical_price[(historical_price.index >= f'{month}-01-{int(year)-5}') & (historical_price.index < f'{month}-01-{year}')]

    weekly_returns = historical_price.pct_change().dropna()
    cov_matrix = weekly_returns.cov()

    symbols = historical_price.columns.to_list()

    results = {}
    stock_list = symbols.copy()

    for stock in stock_list:
        results[stock] = regress_stock_with_index(weekly_returns, stock, nifty50_temp_wr)

    alpha_beta_dict = {}
    alpha_list = []
    beta_list = []

    for stock in stock_list:
        alpha = results[stock]['const']
        beta = results[stock][('Nifty 50',)]
        alpha_list.append(alpha)
        beta_list.append(beta)
        alpha_beta_tuple = (alpha, beta)
        alpha_beta_dict[stock] = alpha_beta_tuple

    index_std_dev = nifty50_temp_wr.std().values[0]
    index_average_return = nifty50_temp_wr.mean().values[0]

    num_symbs = len(historical_price.columns)

    results_dict = smi_max_returns(index_average_return, index_std_dev, alpha_list, beta_list, maximum_risk, num_symbs)

    results_stocks_intersection_of_both[year_month] = results_dict

    portfolio = []
    for i in range(len(symbols)):
        symbol = symbols[i]
        weight = results_dict['Optimal Weights'][i]
        portfolio.append((symbol, weight))

    portfolio_intersection_of_both[year_month] = portfolio

portfolio_intersection_of_both

json_file_path = 'portfolio_intersection_of_both.json'

with open(json_file_path, 'w') as json_file:
    json.dump(portfolio_intersection_of_both, json_file)

json_file_path = 'results_stocks_intersection_of_both.json'

with open(json_file_path, 'w') as json_file:
    json.dump(results_stocks_intersection_of_both, json_file)

json_file_path = 'results_stocks_no_zero_weightage.json'

with open(json_file_path, 'w') as json_file:
    json.dump(results_stocks_no_zero_weightage, json_file)

json_file_path = 'results_stocks_overall_highest_weightage.json'

with open(json_file_path, 'w') as json_file:
    json.dump(results_stocks_overall_highest_weightage, json_file)

results_stocks_overall_highest_weightage

annual_return = 0
annual_variance = 0
annual_std_dev = 0

for year_month, stat_dict in results_stocks_overall_highest_weightage.items():
    annual_return += stat_dict['Annual Return']
    annual_variance += stat_dict['Min Variance']*52
    annual_std_dev += stat_dict['Annual Std Dev']

avg_annual_return = annual_return/61
avg_annual_variance = annual_variance/61
avg_annual_std_dev = annual_std_dev/61

avg_annual_return, avg_annual_variance, avg_annual_std_dev

annual_return = 0
annual_variance = 0
annual_std_dev = 0

for year_month, stat_dict in results_stocks_no_zero_weightage.items():
    annual_return += stat_dict['Annual Return']
    annual_variance += stat_dict['Min Variance']*52
    annual_std_dev += stat_dict['Annual Std Dev']

avg_annual_return = annual_return/61
avg_annual_variance = annual_variance/61
avg_annual_std_dev = annual_std_dev/61

avg_annual_return, avg_annual_variance, avg_annual_std_dev

annual_return = 0
annual_variance = 0
annual_std_dev = 0

for year_month, stat_dict in results_stocks_intersection_of_both.items():
    annual_return += stat_dict['Annual Return']
    annual_variance += stat_dict['Min Variance']*52
    annual_std_dev += stat_dict['Annual Std Dev']

avg_annual_return = annual_return/61
avg_annual_variance = annual_variance/61
avg_annual_std_dev = annual_std_dev/61

avg_annual_return, avg_annual_variance, avg_annual_std_dev
